Refactor blame analysis to new class and clean up code more
